pyproject.toml
[project]
name = "reranking-rag-sys"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cohere>=6.0.0",
    "docling>=2.86.0",
    "fastapi>=0.135.3",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.28",
    "langchain-google-genai>=4.2.1",
    "langchain-postgres>=0.0.17",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.6",
    "pydantic>=2.12.5",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "streamlit>=1.56.0",
    "uvicorn>=0.44.0",
]
